requirements.txt
torch==2.1.0+cpu
torchvision==0.16.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
streamlit
Pillow
numpy
opencv-python